refactor(storage): reorganize conversions from JSON
